requirements.txt
anthropic>=0.40.0
playwright>=1.44.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
python-telegram-bot==20.7
apscheduler>=3.10.0
